quinn

Trait AsyncTimer

pub trait AsyncTimer:
    Send
    + Debug
    + 'static {
    // Required methods
    fn reset(self: Pin<&mut Self>, i: Instant);
    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<()>;
}

Abstract implementation of an async timer for runtime independence

Required Methods

fn reset(self: Pin<&mut Self>, i: Instant)

Update the timer to expire at i

fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<()>

Check whether the timer has expired, and register to be woken if not