quinn

Struct SendStream

pub struct SendStream { /* private fields */ }

A stream that can only be used to send data

If dropped, streams that haven’t been explicitly reset() will continue to (re)transmit previously written data until it has been fully acknowledged or the connection is closed.

Implementations

impl SendStream

pub async fn write(&mut self, buf: &[u8]) -> Result<usize, WriteError>

Write bytes to the stream

Yields the number of bytes written on success. Congestion and flow control may cause this to be shorter than buf.len(), indicating that only a prefix of buf was written.

pub async fn write_all(&mut self, buf: &[u8]) -> Result<(), WriteError>

Convenience method to write an entire buffer to the stream

pub async fn write_chunks( &mut self, bufs: &mut [Bytes], ) -> Result<Written, WriteError>

Write chunks to the stream

Yields the number of bytes and chunks written on success. Congestion and flow control may cause this to be shorter than buf.len(), indicating that only a prefix of bufs was written

pub async fn write_chunk(&mut self, buf: Bytes) -> Result<(), WriteError>

Convenience method to write a single chunk in its entirety to the stream

pub async fn write_all_chunks( &mut self, bufs: &mut [Bytes], ) -> Result<(), WriteError>

Convenience method to write an entire list of chunks to the stream

pub async fn finish(&mut self) -> Result<(), WriteError>

Shut down the send stream gracefully.

No new data may be written after calling this method. Completes when the peer has acknowledged all sent data, retransmitting data as needed.

pub fn reset(&mut self, error_code: VarInt) -> Result<(), UnknownStream>

Close the send stream immediately.

No new data can be written after calling this method. Locally buffered data is dropped, and previously transmitted data will no longer be retransmitted if lost. If an attempt has already been made to finish the stream, the peer may still receive all written data.

pub fn set_priority(&self, priority: i32) -> Result<(), UnknownStream>

Set the priority of the send stream

Every send stream has an initial priority of 0. Locally buffered data from streams with higher priority will be transmitted before data from streams with lower priority. Changing the priority of a stream with pending data may only take effect after that data has been transmitted. Using many different priority levels per connection may have a negative impact on performance.

pub fn priority(&self) -> Result<i32, UnknownStream>

Get the priority of the send stream

pub async fn stopped(&mut self) -> Result<VarInt, StoppedError>

Completes if/when the peer stops the stream, yielding the error code

pub fn id(&self) -> StreamId

Get the identity of this stream